+++ pkg/symbol/addr2line/dwarf.go
@@ -14,6 +14,7 @@
 package addr2line
 
 import (
+	"debug/elf"
 	"fmt"
 	"runtime/debug"
 
@@ -25,25 +26,27 @@ import (
 	"github.com/parca-dev/parca/pkg/symbol/elfutils"
 )
 
+// DwarfLiner is a symbolizer that uses DWARF debug info to symbolize addresses.
 type DwarfLiner struct {
 	logger log.Logger
 
 	dbgFile elfutils.DebugInfoFile
 }
 
-// DWARF is a symbolizer that uses DWARF debug info to symbolize addresses.
-func DWARF(logger log.Logger, path string, demangler *demangle.Demangler) (*DwarfLiner, error) {
-	dbgFile, err := elfutils.NewDebugInfoFile(path, demangler)
+// DWARF creates a new DwarfLiner.
+func DWARF(logger log.Logger, f *elf.File, demangler *demangle.Demangler) (*DwarfLiner, error) {
+	dbgFile, err := elfutils.NewDebugInfoFile(f, demangler)
 	if err != nil {
 		return nil, err
 	}
 
 	return &DwarfLiner{
-		logger:  log.With(logger, "liner", "dwarf", "file", path),
+		logger:  log.With(logger, "liner", "dwarf"),
 		dbgFile: dbgFile,
 	}, nil
 }
 
+// PCToLines returns the resolved source lines for a program counter (memory address).
 func (dl *DwarfLiner) PCToLines(addr uint64) (lines []profile.LocationLine, err error) {
 	defer func() {
 		if r := recover(); r != nil {