+++ ui/packages/shared/profile/src/ProfileView.tsx
@@ -20,7 +20,6 @@ import {
   Button,
   Card,
   SearchNodes,
-  useGrpcMetadata,
   useParcaTheme,
   Callgraph as CallgraphComponent,
 } from '@parca/components';
@@ -31,7 +30,6 @@ import ProfileIcicleGraph from './ProfileIcicleGraph';
 import {ProfileSource} from './ProfileSource';
 import TopTable from './TopTable';
 import useDelayedLoader from './useDelayedLoader';
-import {downloadPprof} from './utils';
 
 import './ProfileView.styles.css';
 
@@ -55,7 +53,7 @@ interface CallgraphData {
   error?: any;
 }
 
-type VisualizationType = 'icicle' | 'table' | 'callgraph' | 'both';
+export type VisualizationType = 'icicle' | 'table' | 'callgraph' | 'both';
 
 interface ProfileVisState {
   currentView: VisualizationType;
@@ -72,6 +70,7 @@ interface ProfileViewProps {
   queryClient?: QueryServiceClient;
   navigateTo?: NavigateFunction;
   compare?: boolean;
+  onDownloadPProf: () => void;
 }
 
 function arrayEquals<T>(a: T[], b: T[]): boolean {
@@ -83,11 +82,18 @@ function arrayEquals<T>(a: T[], b: T[]): boolean {
   );
 }
 export const useProfileVisState = (): ProfileVisState => {
-  const router = parseParams(window.location.search);
-  const currentViewFromURL = router.currentProfileView as string;
-  const [currentView, setCurrentView] = useState<VisualizationType>(
-    (currentViewFromURL as VisualizationType) ?? 'icicle'
-  );
+  const [currentView, setCurrentView] = useState<VisualizationType>(() => {
+    if (typeof window === 'undefined') {
+      return 'icicle';
+    }
+    const router = parseParams(window.location.search);
+    const currentViewFromURL = router.currentProfileView as string;
+
+    if (currentViewFromURL != null) {
+      return currentViewFromURL as VisualizationType;
+    }
+    return 'icicle';
+  });
 
   return {currentView, setCurrentView};
 };
@@ -101,6 +107,7 @@ export const ProfileView = ({
   queryClient,
   navigateTo,
   profileVisState,
+  onDownloadPProf,
 }: ProfileViewProps): JSX.Element => {
   const {ref, dimensions} = useContainerDimensions();
   const [curPath, setCurPath] = useState<string[]>([]);
@@ -108,7 +115,6 @@ export const ProfileView = ({
 
   const [callgraphEnabled] = useUIFeatureFlag('callgraph');
 
-  const metadata = useGrpcMetadata();
   const {loader} = useParcaTheme();
 
   useEffect(() => {
@@ -147,23 +153,6 @@ export const ProfileView = ({
     );
   }
 
-  const downloadPProf = async (e: React.MouseEvent<HTMLElement>): Promise<void> => {
-    e.preventDefault();
-    if (profileSource == null || queryClient == null) {
-      return;
-    }
-
-    try {
-      const blob = await downloadPprof(profileSource.QueryRequest(), queryClient, metadata);
-      const link = document.createElement('a');
-      link.href = window.URL.createObjectURL(blob);
-      link.download = 'profile.pb.gz';
-      link.click();
-    } catch (error) {
-      console.error('Error while querying', error);
-    }
-  };
-
   const resetIcicleGraph = (): void => setCurPath([]);
 
   const setNewCurPath = (path: string[]): void => {
@@ -176,11 +165,12 @@ export const ProfileView = ({
     if (view == null) {
       return;
     }
-    if (navigateTo === undefined) return;
-
     setCurrentView(view);
-    const router = parseParams(window.location.search);
 
+    if (navigateTo === undefined) {
+      return;
+    }
+    const router = parseParams(window.location.search);
     navigateTo('/', {...router, ...{currentProfileView: view}});
   };
 
@@ -199,7 +189,13 @@ export const ProfileView = ({
                     />
                   ) : null}
 
-                  <Button color="neutral" onClick={void downloadPProf}>
+                  <Button
+                    color="neutral"
+                    onClick={e => {
+                      e.preventDefault();
+                      onDownloadPProf();
+                    }}
+                  >
                     Download pprof
                   </Button>
                 </div>